Perubahan kasir dan pembayaran

+++ database/migrations/2025_09_08_042826_create_sale_items_table.php
@@ -6,18 +6,29 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+    /**
+     * Run the migrations.
+     */
     public function up(): void
     {
         Schema::create('sale_items', function (Blueprint $table) {
             $table->id();
+            
+            // Relasi ke tabel 'sales'
             $table->foreignId('sale_id')->constrained()->onDelete('cascade');
+            
+            // Relasi ke tabel 'products'
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
+            
             $table->integer('quantity');
-            $table->integer('price_per_item');
+            $table->decimal('price_per_item', 10, 2); // Mengubah tipe data menjadi decimal
             $table->timestamps();
         });
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down(): void
     {
         Schema::dropIfExists('sale_items');
